@@ -1,10 +1,5 @@
-﻿// Copyright 2020 Connor Andrew Ngo
-// Licensed under the MIT License
-
-namespace Auios.QuadTree
-{
-    public interface IQuadTreeObjectBounds<in T>
-    {
+﻿namespace Auios.QuadTree {
+    public interface IQuadTreeObjectBounds<in T> {
         float GetTop(T obj);
         float GetBottom(T obj);
         float GetLeft(T obj);
